Moved toolbox services,models and data from shared to the toolbox folder
