+++ src/app/feature/toolbox/data/toolbox-topic-metadata.ts
@@ -1,4 +1,4 @@
-import { ToolboxTopicMetadata } from './toolbox.model';
+import { ToolboxTopicMetadata } from '../models/toolbox.model';
 
 export const toolboxTopicNames: ToolboxTopicMetadata[] = [
     {
