+++ src/app/feature/toolbox/toolbox-topic/toolbox-topic.page.ts
@@ -2,8 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, NavigationStart, Router } from '@angular/router';
 import { title } from 'process';
 import { LocalNotificationService } from 'src/app/shared/services/notification/local-notification.service';
-import { ToolboxTopic, ToolboxTopicType, ToolboxSection, ToolboxTip } from 'src/app/shared/services/toolbox/toolbox.model';
-import { ToolboxService } from 'src/app/shared/services/toolbox/toolbox.service';
+import { ToolboxTopic, ToolboxTopicType, ToolboxSection, ToolboxTip } from 'src/app/feature/toolbox/models/toolbox.model';
+import { ToolboxService } from 'src/app/feature/toolbox/services/toolbox.service';
 
 @Component({
   selector: 'plh-toolbox-topic',
