+++ src/app/feature/toolbox/toolbox.page.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { ToolboxTopic, ToolboxTopicMetadata, ToolboxTopicType } from 'src/app/shared/services/toolbox/toolbox.model';
-import { ToolboxService } from 'src/app/shared/services/toolbox/toolbox.service';
+import { ToolboxTopic, ToolboxTopicMetadata, ToolboxTopicType } from 'src/app/feature/toolbox/models/toolbox.model';
+import { ToolboxService } from 'src/app/feature/toolbox/services/toolbox.service';
 import { ThemeService } from '../theme/theme-service/theme.service';
 
 @Component({
